+++ packages/core-cairo/src/build-generic.ts
@@ -2,12 +2,14 @@ import { ERC20Options, buildERC20 } from './erc20';
 import { ERC721Options, buildERC721 } from './erc721';
 import { ERC1155Options, buildERC1155 } from './erc1155';
 import { CustomOptions, buildCustom } from './custom';
+import { AccountOptions, buildAccount } from './account';
 
 export interface KindedOptions {
   ERC20:    { kind: 'ERC20' }    & ERC20Options;
   ERC721:   { kind: 'ERC721' }   & ERC721Options;
   ERC1155:  { kind: 'ERC1155' }  & ERC1155Options;
-  Custom:  { kind: 'Custom' }  & CustomOptions;
+  Account:  { kind: 'Account' }  & AccountOptions;
+  Custom:   { kind: 'Custom' }   & CustomOptions;
 }
 
 export type GenericOptions = KindedOptions[keyof KindedOptions];
@@ -22,6 +24,9 @@ export function buildGeneric(opts: GenericOptions) {
 
     case 'ERC1155':
       return buildERC1155(opts);
+
+    case 'Account':
+      return buildAccount(opts);
 
     case 'Custom':
       return buildCustom(opts);