+++ packages/ui/src/cairo/App.svelte
@@ -7,6 +7,7 @@
     import ERC721Controls from './ERC721Controls.svelte';
     import ERC1155Controls from './ERC1155Controls.svelte';
     import CustomControls from './CustomControls.svelte';
+    import AccountControls from './AccountControls.svelte';
     import CopyIcon from '../icons/CopyIcon.svelte';
     import CheckIcon from '../icons/CheckIcon.svelte';
     import DownloadIcon from '../icons/DownloadIcon.svelte';
@@ -51,6 +52,8 @@
               opts.premint = initialOpts.premint ?? opts.premint;
             case 'ERC721':
               opts.symbol = initialOpts.symbol ?? opts.symbol;
+              break;
+            case 'Account':
               break;
             case 'ERC1155':
             case 'Custom':
@@ -111,6 +114,9 @@
         <button class:selected={tab === 'ERC1155'} on:click={() => tab = 'ERC1155'}>
           ERC1155
         </button>
+        <button class:selected={tab === 'Account'} on:click={() => tab = 'Account'}>
+          Account
+        </button>
         <button class:selected={tab === 'Custom'} on:click={() => tab = 'Custom'}>
           Custom
         </button>
@@ -155,6 +161,9 @@
       </div>
       <div class:hidden={tab !== 'ERC1155'}>
         <ERC1155Controls bind:opts={allOpts.ERC1155}/>
+      </div>
+      <div class:hidden={tab !== 'Account'}>
+        <AccountControls bind:opts={allOpts.Account} errors={errors.Account} accountType={allOpts.Account?.type}/>
       </div>
       <div class:hidden={tab !== 'Custom'}>
         <CustomControls bind:opts={allOpts.Custom} errors={errors.Custom}/>